Change reports to write using streams.
